Moving utils function to analyzer utils.
